List of ProductItems on Product Update screen

+++ app/models/ProductKind.scala
@@ -16,6 +16,6 @@ object ProductKind extends LowPriorityWriteInstances {
       productKindWrites.writes(o) ++ implicitly[OWrites[sorm.Persisted]].writes(o)
   }
 
-  def getById(id: Int) = Db.query[ProductKind].whereEqual("id", id).fetchOne()
+  def getById(id: Long) = Db.query[ProductKind].whereEqual("id", id).fetchOne()
 
 }
